Move model to model/database
